+++ src/i18n/languages/frFR.ts
@@ -101,6 +101,7 @@ const dictionary: RawTranslationPackage = {
     HEX2OCT: 'HEXOCT',
     HLOOKUP: 'RECHERCHEH',
     HOUR: 'HEURE',
+    HYPERLINK: 'LIEN_HYPERTEXTE',
     IF: 'SI',
     IFERROR: 'SIERREUR',
     IFNA: 'SI.NON.DISP',