+++ CHANGELOG.md
@@ -9,8 +9,17 @@ and this project adheres to
 ## [Unreleased]
 
 ### Added
+
+- Added HYPERLINK function.
+  [#1215](https://github.com/handsontable/hyperformula/issues/1215)
+
 - Added IFS function.
   [#1157](https://github.com/handsontable/hyperformula/issues/1157)
+
+### Fixed
+
+- Fixed the `SEARCH` function to be case-insensitive regardless of the engine configuration.
+  [#1225](https://github.com/handsontable/hyperformula/issues/1225)
 
 ## [2.4.0] - 2023-04-24